+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Application/Commands/Query/User/UserOverviewItemDto.cs
@@ -21,7 +21,8 @@ namespace Energinet.DataHub.MarketParticipant.Application.Commands.Query.User;
 public sealed record UserOverviewItemDto(
     Guid Id,
     UserStatus Status,
-    string Name,
+    string FirstName,
+    string LastName,
     string Email,
     string? PhoneNumber,
     DateTimeOffset CreatedDate,