+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Domain/Model/Users/User.cs
@@ -20,8 +20,12 @@ namespace Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 
 public sealed class User
 {
-    public User(ExternalUserId externalId)
+    private const int UserInvitationExpiresAtHours = 24;
+    private readonly SharedUserReferenceId? _sharedId;
+
+    public User(SharedUserReferenceId sharedId, ExternalUserId externalId)
     {
+        _sharedId = sharedId;
         Id = new UserId(Guid.Empty);
         ExternalId = externalId;
         RoleAssignments = new HashSet<UserRoleAssignment>();
@@ -31,21 +35,41 @@ public sealed class User
         UserId id,
         ExternalUserId externalId,
         IEnumerable<UserRoleAssignment> roleAssignments,
-        DateTimeOffset? mitIdSignupInitiatedAt)
+        DateTimeOffset? mitIdSignupInitiatedAt,
+        DateTimeOffset? invitationExpiresAt)
     {
+        _sharedId = null;
         Id = id;
         ExternalId = externalId;
         RoleAssignments = roleAssignments.ToHashSet();
         MitIdSignupInitiatedAt = mitIdSignupInitiatedAt;
+        InvitationExpiresAt = invitationExpiresAt;
     }
 
     public UserId Id { get; }
     public ExternalUserId ExternalId { get; }
+    public SharedUserReferenceId SharedId => _sharedId ?? throw new InvalidOperationException("The shared reference id is only available when creating the entity.");
     public ICollection<UserRoleAssignment> RoleAssignments { get; }
-    public DateTimeOffset? MitIdSignupInitiatedAt { get; private set;  }
-
+    public DateTimeOffset? MitIdSignupInitiatedAt { get; private set; }
+    public DateTimeOffset? InvitationExpiresAt { get; private set;  }
+    public bool ValidLogonRequirements => !InvitationExpiresAt.HasValue || InvitationExpiresAt >= DateTimeOffset.UtcNow;
     public void InitiateMitIdSignup()
     {
         MitIdSignupInitiatedAt = DateTimeOffset.UtcNow;
+    }
+
+    public void EnableUserExpiration()
+    {
+        InvitationExpiresAt = DateTimeOffset.UtcNow.AddHours(UserInvitationExpiresAtHours);
+    }
+
+    public void DisableUserExpiration()
+    {
+        InvitationExpiresAt = null;
+    }
+
+    public bool UserInvitationExpired()
+    {
+        return InvitationExpiresAt.HasValue && InvitationExpiresAt < DateTimeOffset.UtcNow;
     }
 }