+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Domain/Services/UserInvitationService.cs
@@ -14,7 +14,10 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
 using System.Threading.Tasks;
+using Energinet.DataHub.MarketParticipant.Domain.Exception;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 using Energinet.DataHub.MarketParticipant.Domain.Repositories;
@@ -29,6 +32,8 @@ public sealed class UserInvitationService : IUserInvitationService
     private readonly IOrganizationDomainValidationService _organizationDomainValidationService;
     private readonly IUserInviteAuditLogEntryRepository _userInviteAuditLogEntryRepository;
     private readonly IUserRoleAssignmentAuditLogEntryRepository _userRoleAssignmentAuditLogEntryRepository;
+    private readonly IUnitOfWorkProvider _unitOfWorkProvider;
+    private readonly IUserStatusCalculator _userStatusCalculator;
 
     public UserInvitationService(
         IUserRepository userRepository,
@@ -36,7 +41,9 @@ public sealed class UserInvitationService : IUserInvitationService
         IEmailEventRepository emailEventRepository,
         IOrganizationDomainValidationService organizationDomainValidationService,
         IUserInviteAuditLogEntryRepository userInviteAuditLogEntryRepository,
-        IUserRoleAssignmentAuditLogEntryRepository userRoleAssignmentAuditLogEntryRepository)
+        IUserRoleAssignmentAuditLogEntryRepository userRoleAssignmentAuditLogEntryRepository,
+        IUnitOfWorkProvider unitOfWorkProvider,
+        IUserStatusCalculator userStatusCalculator)
     {
         _userRepository = userRepository;
         _userIdentityRepository = userIdentityRepository;
@@ -44,6 +51,8 @@ public sealed class UserInvitationService : IUserInvitationService
         _organizationDomainValidationService = organizationDomainValidationService;
         _userInviteAuditLogEntryRepository = userInviteAuditLogEntryRepository;
         _userRoleAssignmentAuditLogEntryRepository = userRoleAssignmentAuditLogEntryRepository;
+        _unitOfWorkProvider = unitOfWorkProvider;
+        _userStatusCalculator = userStatusCalculator;
     }
 
     public async Task InviteUserAsync(UserInvitation invitation, UserId invitationSentByUserId)
@@ -57,7 +66,10 @@ public sealed class UserInvitationService : IUserInvitationService
         var invitedUser = await GetUserAsync(invitation.Email).ConfigureAwait(false);
         if (invitedUser == null)
         {
+            var sharedId = new SharedUserReferenceId();
+
             var userIdentity = new UserIdentity(
+                sharedId,
                 invitation.Email,
                 invitation.FirstName,
                 invitation.LastName,
@@ -68,7 +80,9 @@ public sealed class UserInvitationService : IUserInvitationService
                 .CreateAsync(userIdentity)
                 .ConfigureAwait(false);
 
-            invitedUser = new User(userIdentityId);
+            invitedUser = new User(sharedId, userIdentityId);
+
+            invitedUser.EnableUserExpiration();
         }
 
         var userInviteRoleAssignments = new List<UserRoleAssignment>();
@@ -80,16 +94,74 @@ public sealed class UserInvitationService : IUserInvitationService
             userInviteRoleAssignments.Add(assignment);
         }
 
-        await _emailEventRepository
-            .InsertAsync(new EmailEvent(invitation.Email, EmailEventType.UserInvite))
+        var uow = await _unitOfWorkProvider
+            .NewUnitOfWorkAsync()
             .ConfigureAwait(false);
 
-        var invitedUserId = await _userRepository
-            .AddOrUpdateAsync(invitedUser)
+        await using (uow.ConfigureAwait(false))
+        {
+            var invitedUserId = await _userRepository
+                .AddOrUpdateAsync(invitedUser)
+                .ConfigureAwait(false);
+
+            await _emailEventRepository
+                .InsertAsync(new EmailEvent(invitation.Email, EmailEventType.UserInvite))
+                .ConfigureAwait(false);
+
+            await AuditLogUserInviteAsync(invitedUserId, invitationSentByUserId, invitation.AssignedActor.Id).ConfigureAwait(false);
+            await AuditLogUserInviteAndUserRoleAssignmentsAsync(invitedUserId, userInviteRoleAssignments, invitationSentByUserId).ConfigureAwait(false);
+
+            await uow.CommitAsync().ConfigureAwait(false);
+        }
+    }
+
+    public async Task ReInviteUserAsync(UserId userId, UserId invitationSentByUserId)
+    {
+        ArgumentNullException.ThrowIfNull(userId);
+
+        var user = await _userRepository.GetAsync(userId).ConfigureAwait(false);
+
+        if (user == null)
+        {
+            throw new NotFoundValidationException($"The specified user {userId} was not found.");
+        }
+
+        var userIdentity = await _userIdentityRepository.GetAsync(user.ExternalId).ConfigureAwait(false);
+
+        if (userIdentity == null)
+        {
+            throw new NotFoundValidationException($"The specified user identity {user.ExternalId} was not found.");
+        }
+
+        if (_userStatusCalculator.CalculateUserStatus(userIdentity.Status, user.InvitationExpiresAt) != UserStatus.InviteExpired)
+        {
+            throw new ValidationException($"The current user invitation for user {userId} is not expired and cannot be re-invited.");
+        }
+
+        user.EnableUserExpiration();
+
+        await _userIdentityRepository.EnableUserAccountAsync(userIdentity.Id).ConfigureAwait(false);
+
+        var assignedActor = user.RoleAssignments.First().ActorId;
+
+        var uow = await _unitOfWorkProvider
+            .NewUnitOfWorkAsync()
             .ConfigureAwait(false);
 
-        await AuditLogUserInviteAsync(invitedUserId, invitationSentByUserId, invitation).ConfigureAwait(false);
-        await AuditLogUserInviteAndUserRoleAssignmentsAsync(invitedUserId, userInviteRoleAssignments, invitationSentByUserId).ConfigureAwait(false);
+        await using (uow.ConfigureAwait(false))
+        {
+            var invitedUserId = await _userRepository
+                .AddOrUpdateAsync(user)
+                .ConfigureAwait(false);
+
+            await _emailEventRepository
+                .InsertAsync(new EmailEvent(userIdentity.Email, EmailEventType.UserInvite))
+                .ConfigureAwait(false);
+
+            await AuditLogUserInviteAsync(invitedUserId, invitationSentByUserId, assignedActor).ConfigureAwait(false);
+
+            await uow.CommitAsync().ConfigureAwait(false);
+        }
     }
 
     private async Task<User?> GetUserAsync(EmailAddress email)
@@ -122,12 +194,12 @@ public sealed class UserInvitationService : IUserInvitationService
         }
     }
 
-    private Task AuditLogUserInviteAsync(UserId toUserId, UserId invitationSentByUserId, UserInvitation invitation)
+    private Task AuditLogUserInviteAsync(UserId toUserId, UserId invitationSentByUserId, ActorId assignedActor)
     {
         var userInviteAuditLog = new UserInviteAuditLogEntry(
             toUserId,
             invitationSentByUserId,
-            invitation.AssignedActor.Id,
+            assignedActor,
             DateTimeOffset.UtcNow);
 
         return _userInviteAuditLogEntryRepository